Ex: 4.21*: Blog List Expansion, step 9

+++ part4/bloglist/controllers/blogs.js
@@ -21,7 +21,7 @@ blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    
+
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
     }
@@ -49,8 +49,30 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    if (!user) {
+        return response.status(400).json({ error: 'userId missing or not valid' })
+    }
+
+    const blogToDelete = await Blog.findById(request.params.id)
+    if (!blogToDelete) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blogToDelete.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }
+    else {
+        return response.status(403).json({ error: 'unauthorized to delete this blog' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
